+++ lib/screens/create_account_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
+
+class CreateAccountScreen extends StatelessWidget {
+  const CreateAccountScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return NeumorphicApp(
+      home: NeumorphicButton(
+        child: NeumorphicText('test'),
+      ),
+    );
+  }
+}
